Fixed some more type conversion issues

+++ Tests/buildBoundary/buildBoundary.cpp
@@ -42,7 +42,7 @@ int main() {
 
     // assert boundary normal vectors are perpendicular to x direction
     auto xplane = rayTriple<NumericType>{1., 0., 0.};
-    for (size_t i = 0; i < 8; i++) {
+    for (unsigned int i = 0; i < 8; i++) {
       auto normal = boundary.getPrimNormal(i);
       RAYTEST_ASSERT_ISNORMAL(normal, xplane, eps)
     }
@@ -68,7 +68,7 @@ int main() {
 
     // assert boundary normal vectors are perpendicular to y direction
     auto yplane = rayTriple<NumericType>{0., 1., 0.};
-    for (size_t i = 0; i < 8; i++) {
+    for (unsigned int i = 0; i < 8; i++) {
       auto normal = boundary.getPrimNormal(i);
       RAYTEST_ASSERT_ISNORMAL(normal, yplane, eps)
     }
@@ -94,7 +94,7 @@ int main() {
 
     // assert boundary normal vectors are perpendicular to x direction
     auto zplane = rayTriple<NumericType>{0., 0., 1.};
-    for (size_t i = 0; i < 8; i++) {
+    for (unsigned int i = 0; i < 8; i++) {
       auto normal = boundary.getPrimNormal(i);
       RAYTEST_ASSERT_ISNORMAL(normal, zplane, eps)
     }
